Implementazione Logica Sfoglia Giochi
Aggiornata la HomePage e attivato il link " Avventura " collegata alla rispettiva Servlet " SfogloaAvventuraGiochiServlet ".
Aggiornati i DAO aggiungendo i metodi per ottenere la lista giochi per genere ID
Creata la cartella private-jsp  e spostata la jsp Sfoglia Giochi Avventura

+++ main/java/org/elis/dao/GiocoDAO.java
@@ -14,7 +14,7 @@ public interface GiocoDAO {
 	Gioco getGiocoByName(String nome);
 	Gioco getGiocoById(long id);
 	List<Gioco> getAllGiochi();
-	List<Gioco> getAllGiochiByGenere(int genere);
+	List<Gioco> getAllGiochiByGenere(long genere);
 	Gioco updateGiocoNome(Gioco gioco, String nome);
 	Gioco updateGiocoImmagine(Gioco gioco, String path);
 	Gioco updateGiocoDataRilascio(Gioco gioco, LocalDate dataRilascio);
